fix rendering of svgs containing line breaks
domPurify was breaking <br/> as can appear in text blocks. This change
adjusts the domPurify settings to fix this and adds a new proisioning
dashboard to demonstrate it.

Note drivable labels still cannot have line breaks, but separate
multi-line text blocks are now supported on the SVG.

Making multi-line labels drivable is doable but it's quite a bit of
rework and there's not an obvious usecase for it.