fix time url forwarding
if a link is configured to forward time it would forward from and to
regardless of whether they were set or no resulting in from & to being
null and the downstream dashboard in a broken state.

This change only forwards time values if they are actually set on the
url.