fix under 100pc scaling
this adjusts the css and html to make the svg correctly
scale in all dimensions and all scales. It's been tested with
svgs with real width/height and viewbox as well as with precentage
width/height and viewbox. With this fixed the default 90% scaling
has been changed to 100% as it now nicely fills the panel.