fix timeslider range wrt timeshift
The incoming panel time-range used to be used as the starting point
for the time-slider range and then extended based off of data. Now
data is used to source the time-range and only if null is the
panel range used. This corrects a problem with query-time-shift which
extends the panels time range to cover both shifted and unshifted
ranges.