slider label
The slider label is set from two pathways with the first not
correctly using the timezone. This has now been made consistent.

The slider label wrapping wasn't disallowed but there isn't enough
vertical space to support wrapping. In addition the animation control
button wasn't fully allowed for and the optional 12 hour 'AM/PM' wasn't
fully allowed for. I explored not passing in width and just going for
100%. Whilst it worked the font isn't constant width so it resulted in
the time slider horizontally wobbling as it was dragged around. So I've
kept the fixed width but left space for 'AM/PM'.